+++ tests/pda-derivation/programs/pda-derivation/src/lib.rs
@@ -3,7 +3,7 @@
 
 use anchor_lang::prelude::*;
 
-declare_id!("ATD5DsyrGhh6bV5PigQ5qQ48uYAChk68YBTkUvdqyJZj");
+declare_id!("Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS");
 
 pub const MY_SEED: [u8; 2] = *b"hi";
 pub const MY_SEED_STR: &str = "hi";